sr_model/requirements.txt
basicsr>=1.3.4.0
facexlib>=0.2.3
lmdb
numpy<1.21  # numba requires numpy<1.21,>=1.17
opencv-python
pyyaml
scipy
torch>=1.7
torchvision
moviepy
tqdm
yapf
sklearn
einops
timm
realesrgan


tb-nightly
lpips
easydict